+++ packages/utils/src/extension.js
@@ -1,6 +1,7 @@
 // import { useCallback } from 'react'
 import { GET_STORAGE, SET_STORAGE } from './constants'
 import { promisifyObject } from './promisify'
+import { pascalize } from './functional'
 /*
   Extension Lib
   -----
@@ -53,18 +54,6 @@ export const alarms = {
   ...chrome.alarms,
   getAlarm,
   upsertAlarm,
-}
-
-export const capitalize = word => word[0].toUpperCase() + word.slice(1)
-
-export const pascalize = key => {
-  const splitKey = key.split('-')
-  if (splitKey.length === 1) {
-    return key
-  }
-  const firstElement = splitKey[0]
-  const transform = splitKey.slice(1).map(value => capitalize(value))
-  return firstElement + transform.join('')
 }
 
 export const useTranslation = () => ({
